# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Platform Infrastructure Orchestrator API
  version: 1.1.0
  description: >-
    Authoritative OpenAPI specification for the Platform Infrastructure Orchestrator,
    governing container stack lifecycles, autoscaling simulation, health verification,
    disaster recovery backups, platform setup bootstrapping, Cloudflare tunnel ingress,
    GDPR right-to-erasure workflows, port diagnostics, and cryptographic certificate lifecycle management.
servers:
  - url: http://localhost:31499/api/v1
    description: Local Platform Orchestrator Daemon
paths:
  /stack/up:
    post:
      summary: Start Infrastructure Stack
      operationId: startStack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StackUpRequest'
      responses:
        '200':
          description: Stack successfully initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackActionResponse'
        '400':
          $ref: '#/components/responses/StandardErrorResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
  /stack/down:
    post:
      summary: Stop Infrastructure Stack
      operationId: stopStack
      responses:
        '200':
          description: Stack successfully stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackActionResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
...
  /scale/service:
    post:
      summary: Scale a Stateless Service
      operationId: scaleService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaleServiceRequest'
      responses:
        '200':
          description: Service replica count updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScaleServiceResponse'
        '400':
          $ref: '#/components/responses/StandardErrorResponse'
  /scale/node:
    post:
      summary: Launch Simulated Compute Node
      operationId: launchComputeNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaunchComputeNodeRequest'
      responses:
        '200':
          description: Compute node successfully launched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaunchComputeNodeResponse'
        '400':
          $ref: '#/components/responses/StandardErrorResponse'
...
  /certs/generate:
    post:
      summary: Generate Platform TLS Certificates
      operationId: generateCerts
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateCertsRequest'
      responses:
        '200':
          description: Certificates successfully generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateCertsResponse'
  /backup/execute:
    post:
      summary: Execute Database Disaster Recovery Backup and Optional Purge
      operationId: executeBackup
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupExecuteRequest'
      responses:
        '200':
          description: Disaster recovery backup outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupExecuteResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
  /setup/bootstrap:
    post:
      summary: Execute End-to-End 7-Step Bootstrapping Pipeline
      operationId: bootstrapPlatform
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupBootstrapRequest'
      responses:
        '200':
          description: Platform setup report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupBootstrapResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
...
  /cloudflare/token:
    post:
      summary: Configure and Save Cloudflare Tunnel Token
      operationId: configureCloudflareToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudflareTokenRequest'
      responses:
        '200':
          description: Token successfully saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudflareActionResponse'
        '400':
          $ref: '#/components/responses/StandardErrorResponse'
  /cloudflare/start:
    post:
      summary: Start Cloudflare Tunnel Ingress Container
      operationId: startCloudflareTunnel
      responses:
        '200':
          description: Cloudflare Tunnel started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudflareActionResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
...
  /gdpr/erasure:
    post:
      summary: Execute GDPR and CCPA Right-to-Erasure Workflow
      operationId: executeGdprErasure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GdprErasureRequest'
      responses:
        '200':
          description: Data erasure audit outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GdprErasureResponse'
        '400':
          $ref: '#/components/responses/StandardErrorResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
...
  /grafana/datasources:
    get:
      summary: List Configured Grafana Datasources
      operationId: listDatasources
      responses:
        '200':
          description: List of datasources returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourcesListResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
    post:
      summary: Create a New Grafana Datasource
      operationId: createDatasource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasourcePayload'
      responses:
        '200':
          description: Datasource created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleDatasourceResponse'
        '400':
          $ref: '#/components/responses/StandardErrorResponse'
  /grafana/datasources/sync:
    post:
      summary: Batch Sync Platform Datasources to Grafana
      operationId: syncDatasources
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasourceSyncOptions'
      responses:
        '200':
          description: Datasource synchronization report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceSyncReportResponse'
...
  /grafana/dashboards:
    get:
      summary: Search Grafana Dashboards
      operationId: searchDashboards
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: folderUids
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Matching dashboards list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardsSearchResponse'
    post:
      summary: Create or Update Dashboard via DB Payload
      operationId: saveDashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardPayload'
      responses:
        '200':
          description: Dashboard save result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSaveResponse'
  /grafana/dashboards/import:
    post:
      summary: Import Dashboard from File or Remote URL
      operationId: importDashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardImportOptions'
      responses:
        '200':
          description: Dashboard imported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSaveResponse'
...
  /grafana/alerts:
    get:
      summary: List Grafana Alert Rules
      operationId: listAlertRules
      responses:
        '200':
          description: Configured alert rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRulesListResponse'
    post:
      summary: Create or Update Grafana Alert Rule
      operationId: saveAlertRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRulePayload'
      responses:
        '200':
          description: Alert rule operation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertOperationResponse'
...
  /grafana/contact-points:
    get:
      summary: List Grafana Alerting Contact Points
      operationId: listContactPoints
      responses:
        '200':
          description: Contact points list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactPointsListResponse'
    post:
      summary: Create or Update Contact Point
      operationId: saveContactPoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactPointPayload'
      responses:
        '200':
          description: Contact point save result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertOperationResponse'
  /grafana/contact-points/test:
    post:
      summary: Test Alert Notification Dispatch to Contact Point
      operationId: testContactPoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactPointPayload'
      responses:
        '200':
          description: Test notification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertOperationResponse'
...
  /services:
    get:
      summary: List Registered External Data Services
      operationId: listExternalServices
      responses:
        '200':
          description: External services list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicesListResponse'
    post:
      summary: Register New External Data Service
      operationId: registerExternalService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceDefinition'
      responses:
        '200':
          description: Service successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailResponse'
...
  /traefik/overview:
    get:
      summary: Get Traefik Ingress Diagnostic Overview
      operationId: getTraefikOverview
      responses:
        '200':
          description: Overview diagnostic report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraefikOverviewResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
...
  /traefik/routers:
    get:
      summary: List Active HTTP Routers
      operationId: listTraefikHTTPRouters
      responses:
        '200':
          description: List of HTTP routers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraefikRoutersListResponse'
    post:
      summary: Register or Update Dynamic HTTP Router
      operationId: saveTraefikHTTPRouter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraefikHTTPRouterDefinition'
      responses:
        '200':
          description: Router registration status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraefikRouterActionResponse'
  /traefik/routers/{name}:
    get:
      summary: Get HTTP Router Details by Name
      operationId: getTraefikHTTPRouter
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HTTP router definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraefikRouterDetailResponse'
        '404':
          $ref: '#/components/responses/StandardErrorResponse'
    delete:
      summary: Remove Dynamic HTTP Router by Name
      operationId: deleteTraefikHTTPRouter
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraefikRouterActionResponse'
...
  /traefik/tcp/routers:
    get:
      summary: List Active TCP/gRPC Routers
      operationId: listTraefikTCPRouters
      responses:
        '200':
          description: List of TCP routers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraefikTCPRoutersListResponse'
    post:
      summary: Register or Update Dynamic TCP Router
      operationId: saveTraefikTCPRouter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraefikTCPRouterDefinition'
      responses:
        '200':
          description: TCP router registration status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraefikRouterActionResponse'
  /traefik/tcp/routers/{name}:
    get:
      summary: Get TCP Router Details by Name
      operationId: getTraefikTCPRouter
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: TCP router definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraefikTCPRouterDetailResponse'
    delete:
      summary: Remove Dynamic TCP Router by Name
      operationId: deleteTraefikTCPRouter
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletion result
          content:
            application/json:
              schema:
  /dns/records:
    get:
      summary: List Discovered Local Platform Domains
      operationId: listDnsRecords
      responses:
        '200':
          description: Discovered DNS records list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DnsRecordsListResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
  /dns/sync:
    post:
      summary: Synchronize Local Platform Domains to Hosts Resolver
      operationId: syncDnsHosts
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DnsSyncRequest'
      responses:
        '200':
          description: Synchronization result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DnsSyncResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
  /dns/check:
    get:
      summary: Verify DNS Domain Resolution Status
      operationId: checkDnsResolution
      responses:
        '200':
          description: Domain resolution diagnostic report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DnsCheckResponse'
        '500':
          $ref: '#/components/responses/StandardErrorResponse'
components:
  responses:
    StandardErrorResponse:
      description: Standard API Error Envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
  schemas:
...
    StackUpRequest:
      type: object
      properties:
        profiles:
          type: array
          items:
            type: string
          example: ["stateless", "compute"]
        detach:
          type: boolean
          default: true
    StackActionResponse:
      type: object
      required:
        - meta
        - data
      properties:
        meta:
          $ref: '#/components/schemas/MetaEnvelope'
        data:
          type: object
          properties:
            message:
              type: string
            status:
              type: string
            activeServices:
              type: array
              items:
                type: string
...
    ScaleServiceRequest:
      type: object
      required:
        - service
        - replicas
      properties:
        service:
          type: string
          example: llmobs-temporal
        replicas:
          type: integer
          minimum: 1
          example: 3
    ScaleServiceResponse:
      type: object
      required:
        - meta
        - data
      properties:
        meta:
          $ref: '#/components/schemas/MetaEnvelope'
        data:
          type: object
          properties:
            service:
              type: string
            replicas:
              type: integer
            message:
              type: string
    LaunchComputeNodeRequest:
      type: object
      required:
        - nodeId
      properties:
        nodeId:
          type: integer
          minimum: 2
          example: 2
        primaryDataHost:
          type: string
          default: host.docker.internal
        enableCloudflare:
          type: boolean
          default: false
    LaunchComputeNodeResponse:
      type: object
      required:
        - meta
        - data
      properties:
        meta:
          $ref: '#/components/schemas/MetaEnvelope'
        data:
          type: object
          properties:
            nodeId:
              type: integer
            projectName:
              type: string
            primaryDataHost:
              type: string
            assignedPorts:
              type: object
              additionalProperties:
                type: integer
...
    GenerateCertsRequest:
      type: object
      properties:
        force:
          type: boolean
          default: false
        domains:
          type: array
          items:
            type: string
    GenerateCertsResponse:
      type: object
      required:
        - meta
        - data
      properties:
        meta:
          $ref: '#/components/schemas/MetaEnvelope'
        data:
          type: object
          properties:
            certPath:
              type: string
            keyPath:
              type: string
            generated:
              type: boolean
    BackupExecuteRequest:
      type: object
      properties:
        backupOnly:
          type: boolean
          default: false
        purge:
          type: boolean
          default: true
        targetDir:
          type: string
...
    SetupBootstrapRequest:
      type: object
      properties:
        pullImages:
          type: boolean
          default: false
...
    CloudflareTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
...
    GdprErasureRequest:
      type: object
      properties:
        userId:
          type: string
        customerId:
          type: string
...